move ./pkg/secrets to ./secrets
